[BUILD_NOTI] Updated configure.ac (static,shared) + Notifications
The configure.ac as altered to allow for both
static and shared compilation of the library.

The behavior can be passed during the autogen.sh-phase:

./autogen.sh --prefix $PWD/disable_static --disable-static
./autogen.sh --prefix $PWD/disable_shared --disable-shared

The Makefile.am was also altered to reflect this configuration
change, see the LTLIBRARIES-changes.

The libev-asynhronous mechanism allows for the passing of a
void* privdata.

The RedisHelper-class now has a function to set this private
pointer.

If none is set the pointer is NULL which means that the old
behavior is preserved.